Move BalancesTable up
